requirements.txt
opencv-python~=4.7.0.72
mediapipe~=0.9.0.1
Pillow~=9.5.0
numpy~=1.21.6